issue#738 address the comments - modify the mADALLogs variable into the application class instead of as locally variable in activities - add the override method of Logger into application - remove the scrolling movement of the log text box
